refactor: main과 컨트롤러 통합
